+++ npl_portfolio/apps.py
@@ -2,4 +2,5 @@ from django.apps import AppConfig
 
 
 class NPLPortfolioConfig(AppConfig):
+    """Class configuring the NPL Portfolio App."""
     name = 'npl_portfolio'