+++ openNPL/sflp_views.py
@@ -26,15 +26,15 @@ from rest_framework.response import Response
 from rest_framework.parsers import JSONParser
 from rest_framework.reverse import reverse
 
-from sflp_portfolio.models.models import Counterparty, Loan, \
-    Enforcement, Forbearance, PropertyCollateral, RepaymentSchedule
+from sflp_portfolio.models.models import Counterparty, CounterpartyState, Loan, \
+    LoanState, Enforcement, Forbearance, PropertyCollateral, PropertyCollateralState
 
 from openNPL.sflp_serializers import SFLP_CounterpartySerializer, SFLP_CounterpartyDetailSerializer
 from openNPL.sflp_serializers import SFLP_EnforcementSerializer, SFLP_EnforcementDetailSerializer
 from openNPL.sflp_serializers import SFLP_ForbearanceSerializer, SFLP_ForbearanceDetailSerializer
 from openNPL.sflp_serializers import SFLP_LoanSerializer, SFLP_LoanDetailSerializer
 from openNPL.sflp_serializers import SFLP_PropertyCollateralSerializer, SFLP_PropertyCollateralDetailSerializer
-from openNPL.sflp_serializers import SFLP_RepaymentScheduleSerializer, SFLP_RepaymentScheduleDetailSerializer
+
 
 @api_view(['GET'])
 @permission_classes((permissions.AllowAny,))
@@ -55,7 +55,7 @@ def api_root(request, format=None):
                 {'sflp_enforcement': reverse('sflp_portfolio:sflp_enforcement_api', request=request, format=format)},
                 {'sflp_forbearance': reverse('sflp_portfolio:sflp_forbearance_api', request=request, format=format)},
                 {'sflp_property_collateral': reverse('sflp_portfolio:sflp_property_collateral_api', request=request,
-                                                    format=format)},
+                                                     format=format)},
                 {'sflp_repayment_schedule': reverse('sflp_portfolio:sflp_repayment_schedule_api', request=request,
                                                     format=format)},
             ]},
@@ -171,7 +171,7 @@ def sflp_property_collateral_api(request):
     if request.method == 'GET':
         property_collateral = PropertyCollateral.objects.all()
         serializer = SFLP_PropertyCollateralSerializer(property_collateral, many=True,
-                                                      context={'request': request})
+                                                       context={'request': request})
         return Response(serializer.data)
     elif request.method == 'POST':
         """
@@ -197,7 +197,6 @@ def sflp_property_collateral_detail(request, pk):
 
     serializer = SFLP_PropertyCollateralDetailSerializer(property_collateral)
     return Response(serializer.data)
-
 
 
 @api_view(['GET', 'POST'])
@@ -272,37 +271,3 @@ def sflp_forbearance_detail(request, pk):
     return Response(serializer.data)
 
 
-@api_view(['GET', 'POST'])
-@permission_classes((permissions.AllowAny,))
-def sflp_repayment_schedule_api(request):
-    """
-    List repayment_schedule (EBA Template)
-    """
-    if request.method == 'GET':
-        repayment_schedule = RepaymentSchedule.objects.all()
-        serializer = SFLP_RepaymentScheduleSerializer(repayment_schedule, many=True, context={'request': request})
-        return Response(serializer.data)
-    elif request.method == 'POST':
-        """
-        Create new repayment_schedule entry
-        """
-        data = JSONParser().parse(request)
-        serializer = SFLP_RepaymentScheduleDetailSerializer(data=data)
-        if serializer.is_valid():
-            serializer.save()
-            return Response(serializer.data, status=status.HTTP_201_CREATED)
-        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
-
-
-@api_view(['GET'])
-def sflp_repayment_schedule_detail(request, pk):
-    """
-    List the data a specific EBA repayment_schedule entry
-    """
-    try:
-        repayment_schedule = RepaymentSchedule.objects.get(pk=pk)
-    except repayment_schedule.DoesNotExist:
-        return Response(status=status.HTTP_404_NOT_FOUND)
-
-    serializer = SFLP_RepaymentScheduleDetailSerializer(repayment_schedule)
-    return Response(serializer.data)