+++ sflp_portfolio/apps.py
@@ -2,4 +2,5 @@ from django.apps import AppConfig
 
 
 class SFLPPortfolioConfig(AppConfig):
+    """Class configuring the SFLP Portfolio App."""
     name = 'sflp_portfolio'