+++ sflp_portfolio/models/enforcement.py
@@ -27,11 +27,8 @@ from sflp_portfolio.models.property_collateral import PropertyCollateral
 
 class Enforcement(models.Model):
     """
-    The Enforcement model holds Enforcement related data
+    The Enforcement model holds Enforcement related data.
 
-     note:: The Agency Single Family Loan Performance Template does not explicitly segment data attributes into Counterparty, Loan etc. The assignment into tables (models) in openNPL is based on the interpretation and main function of different data fields
-
-     .. note:: Fields are currently segmented into static and dynamic. In the future dynamic attributes may move to new models. The distinction is not always clear and may depend on the availability of updated date
 
     """
 
@@ -46,9 +43,10 @@ class Enforcement(models.Model):
     property_collateral_identifier = models.ForeignKey(PropertyCollateral, on_delete=models.CASCADE, null=True,
                                                        blank=True)
 
-    # Loan ID Foreign Key
+
     loan_id = models.ForeignKey(Loan, on_delete=models.CASCADE, blank=True, null=True,
-                                help_text="The load ID to which the Enforcement action belongs")
+                                help_text='The loan ID to which the Enforcement activity links.<a class="risk_manual_url" href="https://www.openriskmanual.org/wiki/FM_SFLP.Loan_Identifier">Documentation</a>')
+    """The loan ID to which the Enforcement activity links."""
 
     #
     # DYNAMIC DATA PROPERTIES