+++ start/views.py
@@ -19,7 +19,6 @@
 # SOFTWARE.
 
 
-from django.shortcuts import render
 from django.views.generic import TemplateView